requirements.txt
# Dependencies for custom implementation
requests==2.31.0
python-dotenv==1.0.0
pycryptodome==3.19.0
urllib3==2.0.7

# Official Alpaca SDK and dependencies
numpy==1.24.3
pandas==2.0.3
alpaca-py==0.8.2
